[update] add pos encoding
